Add serializer for User model
The User model might me swapped out for another and the
intent of the application is to be as reusable as possible. To
create this switch in a clean way get_user_model Django utility
is used.

The serializer assumes that if a custom User model is used, then
it extends from django.contrib.auth.models.CustomUser, which
will always have a USERNAME_FIELD attribute, get_full_name
metod, and is_active attribute.