Add links section to each serializer
The links sections contain a link to itself in the response body.

get_links from django-rest-framework is used. Unlike
Django's reverse, this will return the full URI, including hostname
and protocol, along with the path. For this, reverse needs the
current request, which is passed into the serializer context by
default when the standard ViewSets is used.