Add message subscription to Tornado server
A new dictionary will be created to store the subscriptions when
the application instance is created. The dictionary will map the
sprint ID to a list of connections. The routes are also constructed
here rather than being passed in when the application is created.

Rather than exposing the underlying subscriptions dictionary, the
application exposes the add_subscriber , remove_subscriber and
get_subscribers methods for manipulating and querying the
available subscribers. This abstraction will make it easier to
refactor later.