Implement message identifier
Because RedisSubscriber doesn't know which SprintHandler
put the message on the channel, a client will receive its own
message back.

Each connection handler generates a unique ID when the
connection is opened.

When a message is broadcast, it is wrapped to annotate its
original sender if there was one.

When a message is received on the channel, it is unwrapped to
get the original message and the sender. If there was no
sender, the message is sent to all subscribers, otherwise, it is
sent to all subscribers other than the original sender.

Now the Tornado server can be scaled beyond a single process
by making use of Redis as a simple message broker.