Improve updates
When a user edits a task that change is submitted to the API.
The Django server then notifies the Tornado server, which tells
all clients that there was an update. When a browser client sees
that there is an update, it fetches the task information from the
API. For the user who submitted the original request, this is a
wasted API call. The other connected clients are all making the
same API call to get the same information. It would be more
efficient to send that information along to the clients through
the websocket in the first place.